backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.9
pydantic==2.6.0
sentence-transformers==2.5.1
PyMuPDF==1.23.26
pytesseract==0.3.10
nltk==3.8.1
spellchecker==0.0.1
opencv-python==4.9.0.80
numpy==1.26.4
scikit-learn==1.4.0
duckduckgo-search==4.4.3
openai==1.12.0
Pillow==10.2.0